# Starlark: reuse positional array in native calls where possible