--- a/src/main/java/net/starlark/java/eval/BuiltinFunction.java
+++ b/src/main/java/net/starlark/java/eval/BuiltinFunction.java
@@ -156,6 +156,13 @@ public final class BuiltinFunction implements StarlarkCallable {
 
     ParamDescriptor[] parameters = desc.getParameters();
 
+    // Fast case: reuse positional as Java argument vector
+    if (desc.getPositionalsCanBeJavaArgumentVector()
+        && positional.length == parameters.length
+        && named.length == 0) {
+      return positional;
+    }
+
     // Allocate argument vector.
     int n = parameters.length;
     if (desc.acceptsExtraArgs()) {