@fmeum 2023-10-07
@brandjon Just curious: With progressing Starlarkification, is there a plan to reevaluate the "exectree" representation of Starlark modules at some point?