@tetromino 2024-01-23
Based on a naive reading of the code, we will now skip the special-cased StringModule behavior below (lines 175-179 in the original version of BuiltinFunction.java), and therefore Starlark string object methods that don't take a Starlark thread (e.g. "foo".upper()`) may misbehave. Can you verify whether or not this is the case, and if it's not, add a comment why things nevertheless work? Do we have sufficient test coverage for this?